changed from int to short
git-svn-id: file:///srv/local/data/cleanRepo/trunk@11531 a0517364-e873-4167-b1c8-416daa5a6b91

+++ matsim/src/main/java/org/matsim/lanes/otfvis/io/OTFLaneReader.java
@@ -48,7 +48,7 @@ public class OTFLaneReader extends OTFDataReader {
 
 	@Override
 	public void readConstData(ByteBuffer in) throws IOException {
-		int isQLinkLanesIndicator = in.getInt();
+		short isQLinkLanesIndicator = in.getShort();
 		if (isQLinkLanesIndicator == 1){
 			this.isQLinkLanesReader = true;
 		}
